aplicacion de buenas practicas y documentacion a diseños

+++ src/main/java/com/example/kunturtatto/service/impl/IDesignServiceImpl.java
@@ -34,12 +34,20 @@ public class IDesignServiceImpl implements IDesignService {
 
     @Override
     public void delete(Long idDesign) {
-        designRepository.deleteById(idDesign);
+        if (designRepository.existsById(idDesign)) {
+            designRepository.deleteById(idDesign);
+        }else{
+            throw new IllegalArgumentException("No se puede actualizar: Diseño no encontrado.");
+        }
     }
 
     @Override
     public Design update(Design design) {
-        return designRepository.save(design);        
+        if (designRepository.existsById(design.getIdDesign())) {
+            return designRepository.save(design);        
+        }else{
+            throw new IllegalArgumentException("No se puede actualizar: Diseño no encontrado.");
+        }
     }
 
     @Override
